requirements.txt
streamlit
pandas
mysql-connector-python
faker
python-dotenv